backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
python-socketio>=5.11,<6
pydantic>=2.9,<3
pydantic-settings>=2.6,<3
python-dotenv>=1.0,<2
mediapipe>=0.10,<1
ultralytics>=8.3,<9
faster-whisper>=1.1,<2
openai>=1.55,<2
opencv-python-headless>=4.10,<5
numpy>=1.26,<3
